+++ aiida/backends/tests/utils/archives.py
@@ -85,7 +85,7 @@ def import_archive(archive, filepath=None, external_module=None):
     :param filepath: str of directories of where to find archive (starting "/"s are irrelevant)
     :param external_module: string with name of external module, where archive can be found
     """
-    from aiida.orm.importexport import import_data
+    from aiida.tools.importexport import import_data
 
     dirpath_archive = get_archive_file(archive, filepath=filepath, external_module=external_module)
