+++ aiida/tools/importexport/__init__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+###########################################################################
+# Copyright (c), The AiiDA team. All rights reserved.                     #
+# This file is part of the AiiDA code.                                    #
+#                                                                         #
+# The code is hosted on GitHub at https://github.com/aiidateam/aiida_core #
+# For further information on the license, see the LICENSE.txt file        #
+# For further information please visit http://www.aiida.net               #
+###########################################################################
+# pylint: disable=wildcard-import,undefined-variable
+"""Provides import/export functionalities."""
+from __future__ import division
+from __future__ import print_function
+from __future__ import absolute_import
+
+from .dbexport import export, export_zip
+from .dbimport import *
+from .config import EXPORT_VERSION
+
+__all__ = ('export', 'export_zip', 'EXPORT_VERSION') + dbimport.__all__
